+++ mobile/android/android-components/components/service/firefox-accounts/src/main/java/mozilla/components/service/fxa/sync/SyncManager.kt
@@ -5,7 +5,6 @@
 package mozilla.components.service.fxa.sync
 
 import mozilla.components.concept.sync.SyncableStore
-import mozilla.components.lib.dataprotect.SecureAbove22Preferences
 import mozilla.components.service.fxa.SyncConfig
 import mozilla.components.service.fxa.SyncEngine
 import mozilla.components.service.fxa.manager.SyncEnginesStorage
@@ -78,7 +77,6 @@ interface SyncStatusObserver {
  */
 object GlobalSyncableStoreProvider {
     private val stores: MutableMap<String, SyncableStore> = mutableMapOf()
-    private var keyStorage: SecureAbove22Preferences? = null
 
     /**
      * Configure an instance of [SyncableStore] for a [SyncEngine] enum.
@@ -88,20 +86,9 @@ object GlobalSyncableStoreProvider {
         stores[storePair.first.nativeName] = storePair.second
     }
 
-    /**
-     * Set an instance of [SecureAbove22Preferences] used for accessing an encryption key for [SyncEngine.Passwords].
-     *
-     * @param ks An instance of [SecureAbove22Preferences].
-     */
-    fun configureKeyStorage(ks: SecureAbove22Preferences) {
-        keyStorage = ks
-    }
-
     internal fun getStore(name: String): SyncableStore? {
         return stores[name]
     }
-
-    internal fun getKeyStorage() = keyStorage
 }
 
 /**